# docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        description: Описание возникшей ошибки...
        type: string
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_paid:
        description: Статус оплаты.
        type: boolean
      order_items:
        description: Связь с OrderItem
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      total_amount:
        description: Общая сумма всех заказов.
        type: number
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Связи
      user_id:
        description: Внешний ключ на продавца.
        type: integer
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order:
        allOf:
        - $ref: '#/definitions/models.Order'
        description: Связи
      order_id:
        description: Внешний ключ на заказ.
        type: integer
      price:
        description: Цена товара в момент продажи.
        type: number
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        description: Внешний ключ на продукт.
        type: integer
      quantity:
        description: Количество товара в заказе.
        type: number
      total:
        description: Общая сумма.
        type: number
    type: object
...
  models.SwagUser:
    properties:
      email:
        description: Адрес электронной почты пользователя...
        type: string
      full_name:
        description: Полное имя пользователя...
        type: string
      password:
        description: Пароль для аутентификации...
        type: string
      username:
        description: Логин для входа в систему...
        type: string
    type: object
...
  models.User:
    properties:
      blocked_at:
        description: Время блокировки продавца...
        type: string
      created_at:
        description: Время создания записи...
        type: string
      deleted_at:
        description: Время удаления записи...
        type: string
      email:
        description: Email продавца...
        type: string
      full_name:
        description: Полное имя продавца...
        type: string
      id:
        description: Уникальный идентификатор продавца...
        type: integer
      is_blocked:
        description: Заблокирован ли продавец...
        type: boolean
      is_deleted:
        description: Флаг удаления записи...
        type: boolean
      orders:
        description: Связи
        items:
          $ref: '#/definitions/models.Order'
        type: array
      password:
        description: Пароль продавца...
        type: string
      password_reset_required:
        description: Флаг обязательной смены пароля после сброса Админом...
        type: boolean
      phone:
        description: Телефон продавца...
        type: string
      role:
        description: 'Роль продавца (например: ''admin'', ''manager'', ''seller'')...'
        type: string
      updated_at:
        description: Время последнего обновления записи...
        type: string
      user_settings:
        allOf:
        - $ref: '#/definitions/models.UserSettings'
        description: Связь с таблицей настроек пользователя...
      username:
        description: Логин продавца...
        type: string
    type: object
  models.UserResponse:
    properties:
      blocked_at:
        description: Время блокировки продавца...
        type: string
      created_at:
        description: Время создания записи...
        type: string
      email:
        description: Email продавца...
        type: string
      full_name:
        description: Полное имя продавца...
        type: string
      id:
        description: Уникальный идентификатор продавца...
        type: integer
      is_blocked:
        description: Заблокирован ли продавец...
        type: boolean
      password_reset_required:
        description: Флаг обязательной смены пароля...
        type: boolean
      phone:
        description: Телефон продавца...
        type: string
      role:
        description: Роль продавца...
        type: string
      username:
        description: Логин продавца...
        type: string
    type: object
  models.UserSettings:
    properties:
      accessibility_options:
        description: Опции доступности.
        type: string
      add_confirmation:
        description: Подтверждение добавления.
        type: boolean
      created_at:
        description: Дата создания настроек.
        type: string
      dark_mode_theme:
        description: Темная тема.
        type: boolean
      delete_confirmation:
        description: Подтверждение удаления.
        type: boolean
      desktop_theme:
        description: Тема рабочего стола.
        type: string
      display_language:
        description: Язык отображения.
        type: string
      email_notifications:
        description: Email уведомления.
        type: boolean
      font:
        description: Шрифт.
        type: string
      font_size:
        description: Размер шрифта.
        type: integer
      id:
        description: Уникальный идентификатор настроек пользователя.
        type: integer
      notification_frequency:
        description: Частота уведомлений.
        type: string
      notification_sound:
        description: Звук уведомлений.
        type: boolean
      update_confirmation:
        description: Подтверждение обновления.
        type: boolean
      updated_at:
        description: Дата последнего обновления настроек.
        type: string
      user_id:
        description: Связь с пользователем.
        type: integer
    type: object
host: localhost:8585
info:
  contact: {}
  description: API Server for eShop Application
  title: eShop API
  version: "1.02"
paths:
...
  /users:
    get:
      description: Get a list of all registered users
      operationId: get-all-users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Register a new user (only Admin can do this)
      operationId: create-user
      parameters:
      - description: User Information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SwagUser'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully!!!
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Register a new user
      tags:
      - users
  /users/{id}:
    get:
      description: Get user information by user ID
      operationId: get-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user information by user ID (Admin only)
      operationId: update-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user by ID
      tags:
      - users
  /users/{id}/block:
    patch:
      description: Block a user by ID (Admin only)
      operationId: block-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User blocked successfully!
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Block user by ID
      tags:
      - users
  /users/{id}/hard:
    delete:
      description: Permanently delete user by ID (Admin only)
      operationId: hard-delete-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User hard deleted successfully!
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Hard delete user by ID
      tags:
      - users
  /users/{id}/reset-password:
    patch:
      consumes:
      - application/json
      description: Reset a user's password (Admin only)
      operationId: reset-password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SwagUser'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully!
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reset user password
      tags:
      - users
  /users/{id}/restore:
    patch:
      description: Restore a soft deleted user by ID (Admin only)
      operationId: restore-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User restored successfully!
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Restore user by ID
      tags:
      - users
  /users/{id}/soft:
    delete:
      description: Soft delete user by ID (Admin only)
      operationId: soft-delete-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User soft deleted successfully!
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Soft delete user by ID
      tags:
      - users
  /users/{id}/unblock:
    patch:
      description: Unblock a user by ID (Admin only)
      operationId: unblock-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User unblocked successfully!
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unblock user by ID
      tags:
      - users
  /users/deleted:
    get:
      description: Get a list of all soft deleted users (Admin only)
      operationId: get-all-deleted-users
      produces:
      - application/json
      responses:
        "200":
          description: List of deleted users
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all deleted users
      tags:
      - users
  /users/password:
    patch:
      consumes:
      - application/json
      description: Change the logged-in user's password
      operationId: change-own-password
      parameters:
      - description: Old and new passwords
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SwagUser'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully!
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Unauthorized password change
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Change own password
      tags:
      - users
  /users/settings:
    get:
      description: Get user settings by user ID (only the owner can get their own
        settings)
      operationId: get-user-settings
      produces:
      - application/json
      responses:
        "200":
          description: User settings
          schema:
            $ref: '#/definitions/models.UserSettings'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User settings not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user settings
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user settings (only the owner can update their own settings)
      operationId: update-user-settings
      parameters:
      - description: Updated user settings
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UserSettings'
      produces:
      - application/json
      responses:
        "200":
          description: User settings updated successfully!!!
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user settings
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"